Fatcow Images and Icons
Added various images and icons from Fatcow web hosting (all icons free)
that we may end up using. If not, we can just get rid of them at the end
anyways.

I'm committing this to the master branch since it is a resource for all
to use, and it will have minimal negative impact.