More CSS, Javascript, HTML updates
Added calendar colors to the newevent.php page, along with some more
javascript to auto-resize the description area textbox. Furthermore, I
added a php page (colors.php) which is interpreted as CSS when the
browser reads it. We'll be using this to change the default color of the
calendar that displays (currently just set to what I pulled from
Google).

Also added various minor touchups.