More work on guests section, javascript for checkboxes, datepicker (with nice css) for dates, and some time conflict recognition
This represents quite a few hours of work on various bits and bobs
around the web page. Time was spent on fleshing out html, css, and
javascript in the guest section. Along with that, a large portion of
time was spent on setting up the datepicker, and getting it nicely
formatted with css (cus it sure is ugly to begin with). I also worked on
some javascript for detecting time conflicts (i.e. the end date can't be
before the start date), but I still can't get it to register when the
hour is updated, only the date... Anyways, that's it for today!