Added javascript to time inputs, a section for validation scripts, and a system for displaying errors
Added javascript to the time inputs so that the duration between them is
maintained (even across days) when the start time is changed (i.e.
changing 1pm in 1pm-2pm to 3pm yields 3pm-4pm).

Also added a javascript file for some code validation. This could also
be used for user input validation as well (i.e. autocorrect).

Additionally, added a section to the header.php file to be used as a
simple display mechanism for showing the user errors. More documentation
on this system is in code comments/ will be added.

Finally, I also updated and added some css for these new things I've
added.