WIP update to google icon images, favicon, new notifications area
Added in the favicon (subject to change) for the page. Also, put a
notifications array for letting the user know info that's not as
intimidating as the warning or error popups. Additionally, currently
working on changing how the webpage displays small images as part of a
larger image. This is being changed from php image manipulations to css
image sprites (still WIP, colors area damaged as part of work).