Minor changes to accommodate database
With the addition of the database to the project plan, I've had to
reorganize, delete, and update some content on the page to align with
that.

Additionally, I've removed guests' ability to modify the event (as we
probably don't want them to be doing this anyways). Plus, I've removed
the "calendar" option under the details section, as we're just going to
have users select what calendar they want to save their event to when
it's sent out.