Database
This is a text file containing SQL that can be used to create the
database I have made for our project on your own servers. Use
localhost/phpmyadmin to create the database by running the SQL.

Currently, the database only contains structure for the tables we'll be
using. It does not contain any data.