Minor bug fixing, new page for events list, css reorganization
Fixed a bug with the repeat section so that the summary should now
update when the start date is changed. Also added a new page (currently
the skeleton of one) for displaying events that a user has created and
stored.

Additionally, I did some CSS reorganization to accommodate the new page
added, so that it wouldn't have a bunch of useless CSS added in when the
page loaded.