Lots of work on advanced settings area
Worked on getting advanced settings area into a more usable state,
including the addition of javascript for some interactivity and
formatting, along with a LOT of html structure for the whole section.

Lots of the new css id's haven't been added to their proper files yet,
but that'll get done next time, for now, signing off!