Testing out a new library, and made a new branc
I'm testing out a new library for some optional features we may (though
unlikely) implement, and since this work is still an experimental WIP, I
created a new branch to keep it separate from the rest of the interface.

For the time being, this is only temporary, and will be deleted after
I've either got it into a working state, or decided that it's not worth
the time and effort.