Removing experimental changes
The feature set I was attempting to develop (fancy blurred background
for use in tutorial) isn't working/is taking longer than I would like,
so I'm retiring it for now to continue working on more important issues.

We may implement this later on if we have extra time (unlikely), or if I
decide to add it in after we finish the project for the class.