Finished advanced settings area, starting work on POSTing to database
I've completed the advanced settings area. Client-side validation is
completed, a new way to save input states has been implemented, and
basically, it's ready to ship, so to speak.

Additional features to add in later would be integration in a potential
tutorial system, and the addition of "?" icons with tooltip pop-ups.
Additionally, I could add in the "title" attribute for each input, but
these aren't critical features.

Finally, I'm starting work on submitting POST data to the database, and
integrating that with both newevent.php and eventlist.php