Lots of work on login page.
Added tons of work to the login page (index.php), and it's nearly in a
completed state. I still need to implement a sign-in popup, pop-ups for
info when typing in passwords and the such, a password strength
analyzer, connection to the database, and "is the user already logged
in" checks. Beyond these things, it's mostly done (and it sure is pretty
lookin', mmm, mmm, mmm!).