Finished user authentication, sign in, register, block page access
Did tons of work to setup the my first usage of the database. Made
secure system (salted hashes, etc.) for creating user accounts, signing
in, signing out, and blocking page access when not signed in.

Also added lots of CSS and HTML to support this development, including a
sign in popup in index.php. Also, index.php still needs a small amount
of work (and more optional work, should I have time to implement it) on
the registration popup, specifically a javascript analyzer of password
strength for the user to let them know if they've made a good/bad
password.

Optional features that can be added include, but are not limited to:
dropdown for "My events" to show most recent events created, tutorial
system (interface-experimental), and maybe (really maybe) a blurred
background for the header (but I'm thinking not).