Massive work on eventlist.php
As the summary implies, I've made massive amounts of progress on the
eventlist.php page. I've finished the styling for the page (though minor
tweaks may work their way in), I've added the main table for displaying
events, and I've hooked it into the database to show all events
registered to a particular user.

Furthermore, the table auto-formats the UTC time stored in the database
into the client's local time. Additionally, I've added a field to
tblevents called nColorid, which - while optional - holds the id number
for the color of the event.

On the whole, there's just a LOT of changes, so it's difficult to list
them all here. However, features that still need to be implemented in
eventlist.php include editing an event (and pulling all relevant info to
newevent.php), deleting and event, and creating an event, all of which
require massive overhauls of newevent.php.