Added ability to delete events from eventlist.php
Like the title says, I've added the ability for users to delete events
from their eventlist that they no longer want. Removes records from both
tblevents and tblusersevents.

One potential issue is that the user could refresh the page and cause
the delete script to re-run, but it won't be able to delete anything, as
the record would have already been deleted. So, this could be looked
into later.