Finished saving events to the database
Like the summary says, I've finished coding the queries and whatnot to
save a new event to the database and properly associate it with a user.

During this time, I've also added some more javascript to newevent.php's
javascript files to fix some bugs with syncing up the repeat and
settings dialog boxes with the main page and eachother.

I'm sure that there'll be bugs that pop up, but for now, it works pretty
well. Next step is setting up the ability to edit an event.