Photoshopped a picture and changed how the events list looks
Did some minor touchups on one of the background pictures since it
looked a bit blurry. Saved the original as the same name with an "O" at
the end.

Also added a little hyphen between the start and end times, along with a
change in the order of how the events list displays. It now shows the
most recently modified event on top, and so on down to the bottom.