Lots of work on reloading events from database
Allowing users to edit their events is nearly complete, (at least on
newevent.php). I've added parsing from the database to the page for all
inputs and fields except the advanced settings section so far, which is
what I'll be doing next.

Additionally, I've still got to test the update query that I wrote on
eventlist.php for when the user re-saves their event. As per usual,
everything seems to be working now, but we'll see what eventually
decides to explode on us.