Changed location of sign-in to header.php, instead of index.php
As per issue #5 , I've changed where the sign-in button is located, from
index.php to it's own file at signin.php under pageassembly. This allow
us to have any page be "log-in-able", with the copy-and-pasting of the
php for authentication to each page.

Originally I didn't have it like this as it would have required this
shuffle that I've coded here, but now that it's done, we have a much
more robust system for logging our users in and out.