Finished loading event data to allow editing
As the summary says, I've finished the code to load event data from the
database into newevent.php to act as an editor. There may be
inconsistencies between loading and saving, but I haven't encountered
any yet, so be sure to report any issues found with it.

Additionally, I've added two new background images, and a "FIND TIMES"
button to newevent.php. Currently, it's just a placeholder, but soon
we'll have it hook up with the optimization algorithm.

Furthermore, I've written all of the POST-ing code for the "SEND" button
on newevent.php, though currently it doesn't do anything as
eventlist.php (the destination of the POST request) isn't set up to
handle the data sent.