Finished user account section (pending testing)
I've completed the entire user account section with options to change
their password, change their email, and delete their entire account. All
commands and queries work, and this bit of work is almost ready to go.

One final piece that I have to work on is to get a popup to appear in
the correct position, but I'm pretty sure it's just not working because
my cache isn't cleared.

Regardless, another thing I worked on is a new compartmentalization
system for common POST receivers. Now, all of the database queries
previously copied and pasted across multiple pages are in their own
separate files. This allows to update all pages simultaneously when
making changes to these key queries.