Finished recursive
finished the combination for loop recursive code to find the right
spots, will give the masterCal a -1 to that spot. it adds on with each
invite acceptance. Count them up and put the total in a throw away
matrix NumCal. then goes through a do/while to find order numbers in a
List, and find/return the spot with biggest number to returnCal with at
least Weekly numbers of meeting times