Condensing
I combined the duration loop check with the other larger matrix searcher
to it is now officially good for testing. I also got rid of some
unnecessary matrix holders/variables, and moved the old functions to the
other page just in case I need it again. I just need to get the php
serial stuff down and it should be good to go. Also shoves the