Bug fixes and user registration for attendee fix
Fixed time being displayed incorrectly on eventlist.php (was in UTC, now
in local). Also, modified some styling on eventlist.php to allow better
horizontal scaling. Finally, I've implemented checks to allow for users
to register for an account with an email that is already taken if there
is a token that allows them to do this (this is for registering users
that were attendees in previous events).