Serialize Coding
Change the input data from reading in a file to php serialize, but I'm
pretty sure its wrong. Its a start, but wrong, and I spent 2 hours
online trying to figure out how to code it. made some other minor
adjustments to the coding and such, but no major changes