Began work on displaying optimization algorithm results/popup
I've decided to go with a popup to both show the results of a search,
and to prompt the user to make a new search (or to deny them the ability
to, based on the current state of the event).

While I can progress further in the design of this part of the
interface, I cannot complete it until I know the format of the output
for blOptiSuggestion.