Added password strength analyzer, signin attempt timeout, work on find times button, bug fixing
I've added a password strength analyzer that uses a calculated entropy
value (not sure if this is real entropy or not, but it's close enough)
to display password strength to user. Additionally, I've added a bit of
code to the sign-in process so that if a user has failed logging in
multiple times, it will lock them out from making further attempts for 1
minute (of course, this can be gotten around by closing the browser,
then reopening it, but for most purposes this is ok).

Additionally, I've continued work on making the popup for the find times
button. I've gotten a basic structure down, but I'll need more info
(sample output from algorithm) to finalize this interface element.

Finally, I did a bit of work on fixing some bugs. One worthy of note is
that I fixed a CSS issue on eventlist.php whereby the description field
wouldn't be sized correctly. It should now scale much better with
smaller screens.