Finished an, "I forgot my password" system, bug fixing
As you may be able to tell, this was quite a bit of work to get this
system running. It's a lot of new protocols and security features to
verify everything every step of the way.

I also implemented a few bug fixes relating to styling, and I moved
phpmailer stuff to a centralized phpmailer_init.php file.

Furthermore, I've added a new field to tbltokens (shouldn't affect
current code) that can be used to store json data associated with a
particular token. Using this system, we can now just send the tokenid
over GET, and get the rest of the info from the new field. In this
system, our raw data isn't being exposed to the users, and thus it is
more secure.

Also, this finishes MP 3.