Working on passing arguments to algorithm
Continuing work on optimize.php to send data from the database to the
algorithm. Decided to move all querying to optimize.php to avoid
incompatibility issues with various mysql libraries for python.

Also worked on fixing various bugs in OptCode.py to get it into a
testable state. Still WIP.