Once again, continuing work on classes
I've done a bunch of bug fixing in the classes to get master matrix
generation done properly. However, it's still a little bit off. I've got
to write a function in Calendar to get the earliest time and earliest
date (along with latest) to properly set CalendarMatrix dimensions for
construct_from_blcalendar.