Finished classes and functions
Added in functions to check if required attendees can attend and to
return list of emails of all available attendees. Also added in a subset
method which returns a subset (from one date to another) of a Matrix.

This commit completes all of the work in classes.py that we need for the
project demonstration.