Marginally increase is_required_attendees_busy calculation time
Removed the necessity to convert a potentially massive string to an int,
and instead just check for the first "1" that the function comes across.
According to profiling, it has improved calculation time by what I would
call a statistically insignificant quantity (0.005 seconds), but hey,
it's just a quick idea that popped into my head.